added transcriber task, changed interface to return array of transcriptions
