+++ Examples/WhisperAX/WhisperAX/Views/ContentView.swift
@@ -1118,12 +1118,12 @@ struct ContentView: View {
                 return
             }
 
-            self.tokensPerSecond = transcription?.timings?.tokensPerSecond ?? 0
-            self.effectiveRealTimeFactor = transcription?.timings?.realTimeFactor ?? 0
-            self.currentEncodingLoops = Int(transcription?.timings?.totalEncodingRuns ?? 0)
-            self.firstTokenTime = transcription?.timings?.firstTokenTime ?? 0
-            self.pipelineStart = transcription?.timings?.pipelineStart ?? 0
-            self.currentLag = transcription?.timings?.decodingLoop ?? 0
+            self.tokensPerSecond = transcription?.timings.tokensPerSecond ?? 0
+            self.effectiveRealTimeFactor = transcription?.timings.realTimeFactor ?? 0
+            self.currentEncodingLoops = Int(transcription?.timings.totalEncodingRuns ?? 0)
+            self.firstTokenTime = transcription?.timings.firstTokenTime ?? 0
+            self.pipelineStart = transcription?.timings.pipelineStart ?? 0
+            self.currentLag = transcription?.timings.decodingLoop ?? 0
 
             self.confirmedSegments = segments
         }
@@ -1178,12 +1178,14 @@ struct ContentView: View {
             if progress.avgLogprob! < options.logProbThreshold! {
                 return false
             }
-
             return nil
         }
 
-        let transcription = try await whisperKit.transcribe(audioArray: samples, decodeOptions: options, callback: decodingCallback)
-        return transcription
+        return try await whisperKit.transcribe(
+            audioArray: samples,
+            decodeOptions: options,
+            callback: decodingCallback
+        ).first
     }
 
     // MARK: Streaming Logic
@@ -1277,14 +1279,14 @@ struct ContentView: View {
             // Run realtime transcribe using word timestamps for segmentation
             let transcription = try await transcribeEagerMode(Array(currentBuffer))
             await MainActor.run {
-                self.tokensPerSecond = transcription?.timings?.tokensPerSecond ?? 0
-                self.firstTokenTime = transcription?.timings?.firstTokenTime ?? 0
-                self.pipelineStart = transcription?.timings?.pipelineStart ?? 0
-                self.currentLag = transcription?.timings?.decodingLoop ?? 0
-                self.currentEncodingLoops = Int(transcription?.timings?.totalEncodingRuns ?? 0)
+                self.tokensPerSecond = transcription?.timings.tokensPerSecond ?? 0
+                self.firstTokenTime = transcription?.timings.firstTokenTime ?? 0
+                self.pipelineStart = transcription?.timings.pipelineStart ?? 0
+                self.currentLag = transcription?.timings.decodingLoop ?? 0
+                self.currentEncodingLoops = Int(transcription?.timings.totalEncodingRuns ?? 0)
 
                 let totalAudio = Double(currentBuffer.count) / Double(WhisperKit.sampleRate)
-                self.totalInferenceTime = transcription?.timings?.fullPipeline ?? 0
+                self.totalInferenceTime = transcription?.timings.fullPipeline ?? 0
                 self.effectiveRealTimeFactor = Double(totalInferenceTime) / totalAudio
             }
         } else {
@@ -1299,14 +1301,14 @@ struct ContentView: View {
                     return
                 }
 
-                self.tokensPerSecond = transcription?.timings?.tokensPerSecond ?? 0
-                self.firstTokenTime = transcription?.timings?.firstTokenTime ?? 0
-                self.pipelineStart = transcription?.timings?.pipelineStart ?? 0
-                self.currentLag = transcription?.timings?.decodingLoop ?? 0
-                self.currentEncodingLoops += Int(transcription?.timings?.totalEncodingRuns ?? 0)
+                self.tokensPerSecond = transcription?.timings.tokensPerSecond ?? 0
+                self.firstTokenTime = transcription?.timings.firstTokenTime ?? 0
+                self.pipelineStart = transcription?.timings.pipelineStart ?? 0
+                self.currentLag = transcription?.timings.decodingLoop ?? 0
+                self.currentEncodingLoops += Int(transcription?.timings.totalEncodingRuns ?? 0)
 
                 let totalAudio = Double(currentBuffer.count) / Double(WhisperKit.sampleRate)
-                self.totalInferenceTime += transcription?.timings?.fullPipeline ?? 0
+                self.totalInferenceTime += transcription?.timings.fullPipeline ?? 0
                 self.effectiveRealTimeFactor = Double(totalInferenceTime) / totalAudio
 
                 // Logic for moving segments to confirmedSegments
@@ -1404,7 +1406,7 @@ struct ContentView: View {
         let lastAgreedTokens = lastAgreedWords.flatMap { $0.tokens }
         streamOptions.prefixTokens = lastAgreedTokens
         do {
-            let transcription: TranscriptionResult? = try await whisperKit.transcribe(audioArray: streamingAudio, decodeOptions: streamOptions, callback: decodingCallback)
+            let transcription: TranscriptionResult? = try await whisperKit.transcribe(audioArray: streamingAudio, decodeOptions: streamOptions, callback: decodingCallback).first
             await MainActor.run {
                 var skipAppend = false
                 if let result = transcription {
