+++ Sources/WhisperKit/Core/AudioStreamTranscriber.swift
@@ -22,7 +22,7 @@ public extension AudioStreamTranscriber {
 public typealias AudioStreamTranscriberCallback = (AudioStreamTranscriber.State, AudioStreamTranscriber.State) -> Void
 
 @available(macOS 13, iOS 16, watchOS 10, visionOS 1, *)
-public typealias AudioStreamTranscribe = ([Float], DecodingOptions?, TranscriptionCallback) async throws -> TranscriptionResult?
+public typealias AudioStreamTranscribe = ([Float], DecodingOptions?, TranscriptionCallback) async throws -> [TranscriptionResult]
 
 /// Responsible for streaming audio from the microphone, processing it, and transcribing it in real-time.
 @available(macOS 13, iOS 16, watchOS 10, visionOS 1, *)
@@ -166,7 +166,7 @@ public actor AudioStreamTranscriber {
 
         state.currentText = ""
         state.unconfirmedText = []
-        guard let segments = transcription?.segments else {
+        guard let segments = transcription.first?.segments else {
             return
         }
 
@@ -197,7 +197,7 @@ public actor AudioStreamTranscriber {
         }
     }
 
-    private func transcribeAudioSamples(_ samples: [Float]) async throws -> TranscriptionResult? {
+    private func transcribeAudioSamples(_ samples: [Float]) async throws -> [TranscriptionResult] {
         var options = decodingOptions
         options.clipTimestamps = [state.lastConfirmedSegmentEndSeconds]
         let checkWindow = compressionCheckWindow
